pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code2llm"
version = "0.5.149"
description = "High-performance Python code flow analysis with optimized TOON format - CFG, DFG, call graphs, and intelligent code queries"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Tom Sapletta", email = "tom@sapletta.com"},
]
keywords = [
    "static-analysis",
    "control-flow",
    "data-flow",
    "call-graph",
    "reverse-engineering",
    "toon-format",
    "code-analysis",
    "ast",
    "optimization",
    "complexity-analysis"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "networkx>=2.6",
    "matplotlib>=3.4",
    "pyyaml>=5.4",
    "numpy>=1.20",
    "jinja2>=3.0",
    "radon>=5.1",
    "astroid>=3.0",
    "code2logic",
    "vulture>=2.10",
    "tiktoken>=0.5",
    "tree-sitter>=0.21",
    "tree-sitter-python>=0.21",
    "tree-sitter-javascript>=0.21",
    "tree-sitter-typescript>=0.21",
    "tree-sitter-go>=0.21",
    "tree-sitter-rust>=0.21",
    "tree-sitter-java>=0.21",
    "tree-sitter-c>=0.21",
    "tree-sitter-cpp>=0.22",
    "tree-sitter-c-sharp>=0.21",
    "tree-sitter-php>=0.22",
    "tree-sitter-ruby>=0.21",
]

[project.license]
text = "Apache-2.0"

[project.optional-dependencies]
dev = [
    "pytest>=6.2",
    "pytest-cov>=2.12",
    "pytest-xdist>=3.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.910",
    "goal>=2.1.218",
    "costs>=0.1.20",
    "pfix>=0.1.60",
]

[project.scripts]
code2llm = "code2llm.cli:main"

[project.urls]
Homepage = "https://github.com/wronai/stts"
Repository = "https://github.com/wronai/stts"
Issues = "https://github.com/wronai/stts/issues"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "0.5.14"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-n auto"  # Run tests in parallel using pytest-xdist
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]

[tool.pfix]
# Self-healing Python configuration
model = "openrouter/qwen/qwen3-coder-next"
auto_apply = true
auto_install_deps = true
auto_restart = false
max_retries = 3
create_backups = false
git_auto_commit = false

[tool.pfix.runtime_todo]
enabled = true
todo_file = "TODO.md"
min_severity = "low"
deduplicate = true

[tool.costs]
# AI Cost tracking configuration
badge = true
update_readme = true
readme_path = "README.md"
default_model = "openrouter/qwen/qwen3-coder-next"
analysis_mode = "byok"
full_history = true
max_commits = 500

# Cost thresholds for badge colors (USD)
badge_color_thresholds = { low = 1.0, medium = 5.0, high = 10.0, critical = 50.0 }
